Separated Header Hero and SM
